Separate distinct rows into separate task
